# backend/biosim_server/api/spec/openapi_3_1_0_generated.yaml
openapi: 3.1.0
info:
  title: biosim-server
  version: 0.10.0
paths:
...
  /projects:
    get:
      tags:
      - Projects
      summary: Search published projects (paginated slim results)
      operationId: list-projects
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: 1-indexed page number.
          default: 1
          title: Page
        description: 1-indexed page number.
      - name: perPage
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Results per page.
          default: 20
          title: Perpage
        description: Results per page.
      - name: filters
        in: query
        required: false
        schema:
          type: string
          description: JSON array of {target, allowable_values}.
          default: ''
          title: Filters
        description: JSON array of {target, allowable_values}.
      - name: searchTerm
        in: query
        required: false
        schema:
          type: string
          description: Free-text search across title/abstract/description.
          default: ''
          title: Searchterm
        description: Free-text search across title/abstract/description.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectStubPage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/reindex:
    post:
      tags:
      - Projects
      summary: Rebuild the platform project search index from source collections
      operationId: reindex-projects
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: integer
                title: Response Reindex-Projects
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/stats:
    get:
      tags:
      - Projects
      summary: Facet counts (tags & categories) for a project search
      operationId: list-project-stats
      parameters:
      - name: filters
        in: query
        required: false
        schema:
          type: string
          description: JSON array of {target, allowable_values}.
          default: ''
          title: Filters
        description: JSON array of {target, allowable_values}.
      - name: searchTerm
        in: query
        required: false
        schema:
          type: string
          description: Free-text search across title/abstract/description.
          default: ''
          title: Searchterm
        description: Free-text search across title/abstract/description.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectQueryStat'
                title: Response List-Project-Stats
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{project_id}/summary:
    get:
      tags:
      - Projects
      summary: Platform-owned project summary
      description: Return a public typed summary; upstream drift is a sanitized gateway
        error.
      operationId: get-project-summary
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSummary'
        '502':
          description: The upstream service failed, was unreachable, or returned an
            invalid summary.
        '504':
          description: Timed out while contacting the upstream service.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{project_id}/page:
    get:
      tags:
      - Projects
      summary: Platform-owned project page aggregation
      operationId: get-project-page
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectsPagePayload'
        '429':
          description: The caller exhausted the page-aggregation budget. Retry after
            the indicated delay.
          headers:
            Retry-After:
              description: Seconds until the current fixed window rolls over.
              schema:
                type: integer
        '502':
          description: The upstream service failed, returned an oversized body, or
            returned an invalid page resource.
        '504':
          description: Timed out while contacting the upstream service.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LabeledIdentifier:
      properties:
        uri:
          anyOf:
          - type: string
          - type: 'null'
          title: Uri
        label:
          anyOf:
          - type: string
          - type: 'null'
          title: Label
      type: object
      title: LabeledIdentifier
...
    PageCurve:
      properties:
        id:
          type: string
          title: Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        xDataGenerator:
          anyOf:
          - type: string
          - $ref: '#/components/schemas/PageGenerator'
          title: Xdatagenerator
        yDataGenerator:
          anyOf:
          - type: string
          - $ref: '#/components/schemas/PageGenerator'
          title: Ydatagenerator
        style:
          anyOf:
          - type: string
          - $ref: '#/components/schemas/PageCurveStyle'
          - type: 'null'
          title: Style
      additionalProperties: false
      type: object
      required:
      - id
      - xDataGenerator
      - yDataGenerator
      title: PageCurve
    PageCurveStyle:
      properties:
        base:
          anyOf:
          - type: string
          - type: 'null'
          title: Base
        line:
          anyOf:
          - $ref: '#/components/schemas/PageLineStyle'
          - type: 'null'
        marker:
          anyOf:
          - $ref: '#/components/schemas/PageMarkerStyle'
          - type: 'null'
      additionalProperties: false
      type: object
      title: PageCurveStyle
    PageDataSet:
      properties:
        id:
          type: string
          title: Id
        label:
          type: string
          title: Label
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      additionalProperties: false
      type: object
      required:
      - id
      - label
      title: PageDataSet
...
    PageFile:
      properties:
        format:
          type: string
          title: Format
        location:
          type: string
          title: Location
        size:
          type: integer
          title: Size
        url:
          type: string
          title: Url
      additionalProperties: false
      type: object
      required:
      - format
      - location
      - size
      - url
      title: PageFile
    PageGenerator:
      properties:
        id:
          type: string
          title: Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      additionalProperties: false
      type: object
      required:
      - id
      title: PageGenerator
    PageIdentifier:
      properties:
        uri:
          anyOf:
          - type: string
          - type: 'null'
          title: Uri
        label:
          type: string
          title: Label
      additionalProperties: false
      type: object
      required:
      - label
      title: PageIdentifier
    PageLineStyle:
      properties:
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
        thickness:
          anyOf:
          - type: number
          - type: 'null'
          title: Thickness
        type:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
      additionalProperties: false
      type: object
      title: PageLineStyle
    PageLinkedIdentifier:
      properties:
        uri:
          anyOf:
          - type: string
          - type: 'null'
          title: Uri
        label:
          type: string
          title: Label
      additionalProperties: false
      type: object
      required:
      - label
      title: PageLinkedIdentifier
    PageMarkerStyle:
      properties:
        fillColor:
          anyOf:
          - type: string
          - type: 'null'
          title: Fillcolor
        lineColor:
          anyOf:
          - type: string
          - type: 'null'
          title: Linecolor
        lineThickness:
          anyOf:
          - type: number
          - type: 'null'
          title: Linethickness
        size:
          anyOf:
          - type: number
          - type: 'null'
          title: Size
        type:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
      additionalProperties: false
      type: object
      title: PageMarkerStyle
    PageOutput:
      properties:
        _type:
          type: string
          title: Type
        id:
          type: string
          title: Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        xScale:
          anyOf:
          - type: string
          - type: 'null'
          title: Xscale
        yScale:
          anyOf:
          - type: string
          - type: 'null'
          title: Yscale
        curves:
          anyOf:
          - items:
              $ref: '#/components/schemas/PageCurve'
            type: array
          - type: 'null'
          title: Curves
        dataSets:
          anyOf:
          - items:
              $ref: '#/components/schemas/PageDataSet'
            type: array
          - type: 'null'
          title: Datasets
      additionalProperties: false
      type: object
      required:
      - _type
      - id
      title: PageOutput
...
    PageProject:
      properties:
        id:
          type: string
          title: Id
        created:
          type: string
          title: Created
        updated:
          type: string
          title: Updated
      additionalProperties: false
      type: object
      required:
      - id
      - created
      - updated
      title: PageProject
    PageProjectRun:
      properties:
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        abstract:
          anyOf:
          - type: string
          - type: 'null'
          title: Abstract
        creators:
          items:
            $ref: '#/components/schemas/PageIdentifier'
          type: array
          title: Creators
        keywords:
          items:
            $ref: '#/components/schemas/PageIdentifier'
          type: array
          title: Keywords
        thumbnails:
          items:
            type: string
          type: array
          title: Thumbnails
        projectSize:
          anyOf:
          - type: integer
          - type: 'null'
          title: Projectsize
        resultsSize:
          anyOf:
          - type: integer
          - type: 'null'
          title: Resultssize
        id:
          type: string
          title: Id
        citations:
          items:
            $ref: '#/components/schemas/PageLinkedIdentifier'
          type: array
          title: Citations
        encodes:
          items:
            $ref: '#/components/schemas/PageLinkedIdentifier'
          type: array
          title: Encodes
        simulator:
          $ref: '#/components/schemas/PageSimulator'
        modelFormats:
          items:
            type: string
          type: array
          title: Modelformats
      additionalProperties: false
      type: object
      required:
      - name
      - creators
      - keywords
      - thumbnails
      - id
      - citations
      - encodes
      - simulator
      - modelFormats
      title: PageProjectRun
...
    PageSimulator:
      properties:
        name:
          type: string
          title: Name
        version:
          type: string
          title: Version
      additionalProperties: false
      type: object
      required:
      - name
      - version
      title: PageSimulator
    PageSpecification:
      properties:
        id:
          type: string
          title: Id
        outputs:
          items:
            $ref: '#/components/schemas/PageOutput'
          type: array
          title: Outputs
      additionalProperties: false
      type: object
      required:
      - id
      - outputs
      title: PageSpecification
...
    ProjectQueryStat:
      properties:
        target:
          type: string
          title: Target
        valueFrequencies:
          items:
            $ref: '#/components/schemas/ValueFrequency'
          type: array
          title: Valuefrequencies
      type: object
      required:
      - target
      title: ProjectQueryStat
      description: Facet counts for a single filterable target (e.g. taxa, modelFormat).
    ProjectStub:
      properties:
        id:
          type: string
          title: Id
        simulationRun:
          type: string
          title: Simulationrun
        created:
          type: string
          title: Created
        updated:
          type: string
          title: Updated
        name:
          type: string
          title: Name
        summary:
          type: string
          title: Summary
        model_format:
          type: string
          title: Model Format
          default: ''
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
      type: object
      required:
      - id
      - simulationRun
      - created
      - updated
      - name
      - summary
      title: ProjectStub
      description: 'Slim, list-optimized project record. Mirrors the frontend ``ProjectStub``.


        ``model_format`` and ``image_url`` require a join beyond ``projectSummary``

        (Specifications / Files) and are best-effort until that join lands; they

        default to empty rather than failing the whole page.'
    ProjectStubPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ProjectStub'
          type: array
          title: Items
        total:
          type: integer
          title: Total
      type: object
      required:
      - items
      - total
      title: ProjectStubPage
      description: A page of results plus the total match count (for the pager).
    ProjectSummary:
      properties:
        id:
          type: string
          title: Id
        created:
          type: string
          title: Created
        updated:
          type: string
          title: Updated
        simulationRun:
          $ref: '#/components/schemas/RunSummary'
      type: object
      required:
      - id
      - created
      - updated
      - simulationRun
      title: ProjectSummary
    ProjectsPagePayload:
      properties:
        project:
          $ref: '#/components/schemas/PageProject'
        simulationRun:
          $ref: '#/components/schemas/PageProjectRun'
        files:
          items:
            $ref: '#/components/schemas/PageFile'
          type: array
          title: Files
        specifications:
          items:
            $ref: '#/components/schemas/PageSpecification'
          type: array
          title: Specifications
      additionalProperties: false
      type: object
      required:
      - project
      - simulationRun
      - files
      - specifications
      title: ProjectsPagePayload
...
    RunExecution:
      properties:
        simulator:
          $ref: '#/components/schemas/SimulatorSummary'
        projectSize:
          anyOf:
          - type: integer
          - type: 'null'
          title: Projectsize
        resultsSize:
          anyOf:
          - type: integer
          - type: 'null'
          title: Resultssize
      type: object
      required:
      - simulator
      title: RunExecution
    RunMetadataSummary:
      properties:
        abstract:
          anyOf:
          - type: string
          - type: 'null'
          title: Abstract
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        thumbnails:
          items:
            type: string
          type: array
          title: Thumbnails
        creators:
          items:
            $ref: '#/components/schemas/LabeledIdentifier'
          type: array
          title: Creators
        keywords:
          items:
            $ref: '#/components/schemas/LabeledIdentifier'
          type: array
          title: Keywords
        citations:
          items:
            $ref: '#/components/schemas/LabeledIdentifier'
          type: array
          title: Citations
        encodes:
          items:
            $ref: '#/components/schemas/LabeledIdentifier'
          type: array
          title: Encodes
      type: object
      title: RunMetadataSummary
...
    RunSummary:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        run:
          $ref: '#/components/schemas/RunExecution'
        metadata:
          items:
            $ref: '#/components/schemas/RunMetadataSummary'
          type: array
          title: Metadata
      type: object
      required:
      - id
      - name
      - run
      title: RunSummary
...
    SimulatorSummary:
      properties:
        name:
          type: string
          title: Name
        version:
          type: string
          title: Version
      type: object
      required:
      - name
      - version
      title: SimulatorSummary
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ValueFrequency:
      properties:
        value:
          type: string
          title: Value
        count:
          type: integer
          title: Count
      type: object
      required:
      - value
      - count
      title: ValueFrequency
      description: One facet value and how many matching projects carry it.
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer